# Optimize gaps for RangeSet containing only one range